Move modules into appropriate directories.
