+++ hasher-matcher-actioner/webapp/src/styles/_app.scss
@@ -1,3 +1,7 @@
+/*
+ * Copyright (c) Facebook, Inc. and its affiliates.All Rights Reserved
+ */
+
 div#root {
   display: flex;
   flex-wrap: nowrap;
@@ -14,11 +18,6 @@ div#root {
 .main {
   height: 100vh;
   overflow-y: auto;
-}
-
-.action-rule-classification-column {
-  word-wrap: break-word;
-  max-width: 250px;
 }
 
 .ion-icon-white {
