+++ hasher-matcher-actioner/webapp/src/styles/_settings.scss
@@ -1,0 +1,8 @@
+/*
+ * Copyright (c) Facebook, Inc. and its affiliates.All Rights Reserved
+ */
+
+.action-rule-classification-column {
+  word-wrap: break-word;
+  max-width: 250px;
+}
